refactor: extract `argus-store` crate
