+++ crates/argus-store/src/lib.rs
@@ -1,5 +1,4 @@
-//! This module provides a concrete implementation of the StateRepository using
-//! SQLite.
+//! SQLite-backed persistence implementation for Argus.
 
 use std::str::FromStr;
 
@@ -35,7 +34,7 @@ impl SqliteStateRepository {
     /// Runs database migrations.
     #[tracing::instrument(skip(self), level = "info")]
     pub async fn run_migrations(&self) -> Result<(), PersistenceError> {
-        sqlx::migrate!("./migrations").run(&self.pool).await.map_err(|e| {
+        sqlx::migrate!("../../migrations").run(&self.pool).await.map_err(|e| {
             tracing::error!(error = %e, "Failed to run database migrations.");
             PersistenceError::MigrationError(e.to_string())
         })?;
