+++ src/abi/repository.rs
@@ -76,7 +76,8 @@ mod tests {
     use argus_core::persistence::traits::AppRepository;
 
     use super::*;
-    use crate::{persistence::sqlite::SqliteStateRepository, test_helpers::erc20_abi_json};
+    use argus_store::SqliteStateRepository;
+    use crate::test_helpers::erc20_abi_json;
 
     const REQUIRED_ERC20_FUNCTIONS: &[&str] = &[
         "transfer",
