--- src/persistence/mod.rs
@@ -1,4 +1,0 @@
-//! This module contains the state management logic for the Argus application.
-
-pub mod sqlite;
-pub use sqlite::SqliteStateRepository;
