+++ src/test_helpers/abi.rs
@@ -6,8 +6,8 @@ use argus_core::persistence::traits::AppRepository;
 
 use crate::{
     abi::{AbiRepository, AbiService},
-    persistence::sqlite::SqliteStateRepository,
 };
+use argus_store::SqliteStateRepository;
 
 /// A simple ABI JSON for testing purposes.
 pub fn erc20_abi_json() -> &'static str {
