+++ tests/persistence.rs
@@ -1,6 +1,7 @@
 //! Integration tests for the persistence layer
 
-use argus::{persistence::sqlite::SqliteStateRepository, test_helpers::ActionBuilder};
+use argus::test_helpers::ActionBuilder;
+use argus_store::SqliteStateRepository;
 use argus_core::{
     models::{NetworkId, action::ActionConfig, monitor::MonitorConfig},
     persistence::traits::AppRepository,
